More behavior corrections for the robot. More behavior corrections for the robot.
Made the robot Go back on its search path when returning after the drop off. By pushing the old goal to the interruption stack

Made a counter for targets. Now can tell how many cubes we see.

So the bringing the cubes to center code is complete garbage. It works half the time. Will have to fix that too now.