Improved pick up and drop off
Killed the center recognition code. We know exactly where center is. As soon as we see any center tags we do not try to find the center by using the center tags but instead just proceed to drop the cube. We simply move forward for three seconds and drop cube. Works very well. Killed code it by just commenting everything out (this one is for you J) and writing my own timed logic because that is all I need in our case.

Reduced the return to search delay. For now  I am unsure why it is even used tbh. It is set to 3 sec. So after drop off has been completed the robot will stay still for 3 sec and move on.

Added code to return to the location of the picked up cube before going to old location restore. It works but the robot takes its sweat time recalibrating back to its old path.  But I do not think it is a problem. Recalibrations usually take about 5 -10 secs driving around and sometimes cubes are found in the process.

Adjusted the pick up threshold. This makes robot not try a pickup until its closer. This adjustment makes robot pick up cube almost perfectly in one try. Problem that might come out of this is that the cube might get stuck in the claw. More testing is needed for this.