Fixes
Fixed the bug which stopped the robot from moving when got a collision while carrying a target. Error was caused because the robot got stuck in a mobility loop since he couldnt move to the center. Fixed by giving the rover its own collision logic when with a cube.

Created a Wall safeguard. We shouldn't hit a wall or any immobile object but if we do hit it more than 4 times then the return goal is changed.

Increased the collision distance to 0.75 meters.

Made rovers do a cool 180 backwards turn on calibrate. Rovers would be really popular in high school. Would get girls. And probably would hate their parents.

Calibrated robots 5 and 6. The once that are on the sides like plebs.